requirements.txt
python-dotenv
google-genai
SpeechRecognition
PyAudio
piper-tts
sounddevice
numpy